Refactor generic_controller out of internal/controllers
