+++ v2/internal/controllers/controller_resources.go
@@ -33,6 +33,7 @@ import (
 	resourcesbeta "github.com/Azure/azure-service-operator/v2/api/resources/v1beta20200601"
 	. "github.com/Azure/azure-service-operator/v2/internal/logging"
 	"github.com/Azure/azure-service-operator/v2/internal/reconcilers/arm"
+	"github.com/Azure/azure-service-operator/v2/internal/reconcilers/generic"
 	mysqlreconciler "github.com/Azure/azure-service-operator/v2/internal/reconcilers/mysql"
 	"github.com/Azure/azure-service-operator/v2/internal/reflecthelpers"
 	"github.com/Azure/azure-service-operator/v2/internal/resolver"
@@ -47,7 +48,7 @@ func GetKnownStorageTypes(
 	armClientFactory arm.ARMClientFactory,
 	kubeClient kubeclient.Client,
 	positiveConditions *conditions.PositiveConditionBuilder,
-	options Options) ([]*registration.StorageType, error) {
+	options generic.Options) ([]*registration.StorageType, error) {
 
 	resourceResolver := resolver.NewResolver(kubeClient)
 	knownStorageTypes, err := getGeneratedStorageTypes(mgr, armClientFactory, kubeClient, resourceResolver, positiveConditions, options)
@@ -94,7 +95,7 @@ func getGeneratedStorageTypes(
 	kubeClient kubeclient.Client,
 	resourceResolver *resolver.Resolver,
 	positiveConditions *conditions.PositiveConditionBuilder,
-	options Options) ([]*registration.StorageType, error) {
+	options generic.Options) ([]*registration.StorageType, error) {
 	knownStorageTypes := getKnownStorageTypes()
 
 	knownStorageTypes = append(
@@ -158,7 +159,7 @@ func augmentWithARMReconciler(
 	kubeClient kubeclient.Client,
 	resourceResolver *resolver.Resolver,
 	positiveConditions *conditions.PositiveConditionBuilder,
-	options Options,
+	options generic.Options,
 	extension genruntime.ResourceExtension,
 	t *registration.StorageType) {
 	t.Reconciler = arm.NewAzureDeploymentReconciler(
