+++ v2/internal/reconcilers/generic/generic_reconciler.go
@@ -3,7 +3,7 @@ Copyright (c) Microsoft Corporation.
 Licensed under the MIT license.
 */
 
-package controllers
+package generic
 
 import (
 	"context"
@@ -15,19 +15,12 @@ import (
 	"github.com/pkg/errors"
 	"golang.org/x/time/rate"
 	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/conversion"
 	"k8s.io/apimachinery/pkg/runtime/schema"
-	kerrors "k8s.io/apimachinery/pkg/util/errors"
 	"k8s.io/client-go/tools/record"
 	"k8s.io/client-go/util/workqueue"
 	ctrl "sigs.k8s.io/controller-runtime"
-	ctrlbuilder "sigs.k8s.io/controller-runtime/pkg/builder"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/client/apiutil"
-	"sigs.k8s.io/controller-runtime/pkg/controller"
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
-	"sigs.k8s.io/controller-runtime/pkg/predicate"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
 	"github.com/Azure/azure-service-operator/v2/internal/config"
@@ -37,10 +30,9 @@ import (
 	"github.com/Azure/azure-service-operator/v2/internal/util/kubeclient"
 	"github.com/Azure/azure-service-operator/v2/pkg/genruntime"
 	"github.com/Azure/azure-service-operator/v2/pkg/genruntime/conditions"
-	"github.com/Azure/azure-service-operator/v2/pkg/genruntime/registration"
 )
 
-const GenericControllerFinalizer = "serviceoperator.azure.com/finalizer"
+const GenericReconcilerFinalizer = "serviceoperator.azure.com/finalizer"
 
 // NamespaceAnnotation defines the annotation name to use when marking
 // a resource with the namespace of the managing operator.
@@ -63,129 +55,6 @@ type GenericReconciler struct {
 }
 
 var _ reconcile.Reconciler = &GenericReconciler{} // GenericReconciler is a reconcile.Reconciler
-
-type Options struct {
-	controller.Options
-
-	// options specific to our controller
-	RequeueIntervalCalculator interval.Calculator
-	Config                    config.Values
-	LoggerFactory             func(obj metav1.Object) logr.Logger
-}
-
-func RegisterWebhooks(mgr ctrl.Manager, objs []client.Object) error {
-	var errs []error
-
-	for _, obj := range objs {
-		if err := registerWebhook(mgr, obj); err != nil {
-			errs = append(errs, err)
-		}
-	}
-
-	return kerrors.NewAggregate(errs)
-}
-
-func registerWebhook(mgr ctrl.Manager, obj client.Object) error {
-	_, err := conversion.EnforcePtr(obj)
-	if err != nil {
-		return errors.Wrap(err, "obj was expected to be ptr but was not")
-	}
-
-	return ctrl.NewWebhookManagedBy(mgr).For(obj).Complete()
-}
-
-func RegisterAll(
-	mgr ctrl.Manager,
-	fieldIndexer client.FieldIndexer,
-	kubeClient kubeclient.Client,
-	positiveConditions *conditions.PositiveConditionBuilder,
-	objs []*registration.StorageType,
-	options Options) error {
-
-	// pre-register any indexes we need
-	for _, obj := range objs {
-		for _, indexer := range obj.Indexes {
-			options.LogConstructor(nil).V(Info).Info("Registering indexer for type", "type", fmt.Sprintf("%T", obj.Obj), "key", indexer.Key)
-			err := fieldIndexer.IndexField(context.Background(), obj.Obj, indexer.Key, indexer.Func)
-			if err != nil {
-				return errors.Wrapf(err, "failed to register indexer for %T, Key: %q", obj.Obj, indexer.Key)
-			}
-		}
-	}
-
-	var errs []error
-	for _, obj := range objs {
-		// TODO: Consider pulling some of the construction of things out of register (gvk, etc), so that we can pass in just
-		// TODO: the applicable extensions rather than a map of all of them
-		if err := register(mgr, kubeClient, positiveConditions, obj, options); err != nil {
-			errs = append(errs, err)
-		}
-	}
-
-	return kerrors.NewAggregate(errs)
-}
-
-func register(
-	mgr ctrl.Manager,
-	kubeClient kubeclient.Client,
-	positiveConditions *conditions.PositiveConditionBuilder,
-	info *registration.StorageType,
-	options Options) error {
-
-	// Use the provided GVK to construct a new runtime object of the desired concrete type.
-	gvk, err := apiutil.GVKForObject(info.Obj, mgr.GetScheme())
-	if err != nil {
-		return errors.Wrapf(err, "creating GVK for obj %T", info)
-	}
-
-	loggerFactory := func(mo genruntime.MetaObject) logr.Logger {
-		result := options.LogConstructor(nil)
-		if options.LoggerFactory != nil {
-			if factoryResult := options.LoggerFactory(mo); factoryResult != (logr.Logger{}) && factoryResult != logr.Discard() {
-				result = factoryResult
-			}
-		}
-
-		return result.WithName(info.Name)
-	}
-	eventRecorder := mgr.GetEventRecorderFor(info.Name)
-
-	options.LogConstructor(nil).V(Status).Info("Registering", "GVK", gvk)
-
-	reconciler := &GenericReconciler{
-		Reconciler:                info.Reconciler,
-		KubeClient:                kubeClient,
-		Config:                    options.Config,
-		LoggerFactory:             loggerFactory,
-		Recorder:                  eventRecorder,
-		GVK:                       gvk,
-		PositiveConditions:        positiveConditions,
-		RequeueIntervalCalculator: options.RequeueIntervalCalculator,
-	}
-
-	// Note: These predicates prevent status updates from triggering a reconcile.
-	// to learn more look at https://pkg.go.dev/sigs.k8s.io/controller-runtime/pkg/predicate#GenerationChangedPredicate
-	filter := predicate.Or(
-		predicate.GenerationChangedPredicate{},
-		reconcilers.ARMReconcilerAnnotationChangedPredicate(options.LogConstructor(nil).WithName(info.Name)))
-
-	builder := ctrl.NewControllerManagedBy(mgr).
-		// Note: These predicates prevent status updates from triggering a reconcile.
-		// to learn more look at https://pkg.go.dev/sigs.k8s.io/controller-runtime/pkg/predicate#GenerationChangedPredicate
-		For(info.Obj, ctrlbuilder.WithPredicates(filter)).
-		WithOptions(options.Options)
-
-	for _, watch := range info.Watches {
-		builder = builder.Watches(watch.Src, watch.MakeEventHandler(kubeClient, options.LogConstructor(nil).WithName(info.Name)))
-	}
-
-	err = builder.Complete(reconciler)
-	if err != nil {
-		return errors.Wrap(err, "unable to build controllers / reconciler")
-	}
-
-	return nil
-}
 
 // Reconcile will take state in K8s and apply it to Azure
 func (gr *GenericReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Result, error) {
@@ -309,7 +178,7 @@ func (gr *GenericReconciler) claimResource(ctx context.Context, log logr.Logger,
 	// we issue a PUT to Azure but the commit of the resource into etcd fails, causing us to have an unset
 	// finalizer and have started resource creation in Azure.
 	log.V(Info).Info("adding finalizer")
-	controllerutil.AddFinalizer(metaObj, GenericControllerFinalizer)
+	controllerutil.AddFinalizer(metaObj, GenericReconcilerFinalizer)
 
 	err = gr.KubeClient.CommitObject(ctx, metaObj)
 	if err != nil {
@@ -321,7 +190,7 @@ func (gr *GenericReconciler) claimResource(ctx context.Context, log logr.Logger,
 }
 
 func (gr *GenericReconciler) needToAddFinalizer(metaObj genruntime.MetaObject) bool {
-	unsetFinalizer := !controllerutil.ContainsFinalizer(metaObj, GenericControllerFinalizer)
+	unsetFinalizer := !controllerutil.ContainsFinalizer(metaObj, GenericReconcilerFinalizer)
 	return unsetFinalizer
 }
 
@@ -348,13 +217,13 @@ func (gr *GenericReconciler) delete(ctx context.Context, log logr.Logger, metaOb
 	reconcilePolicy := reconcilers.GetReconcilePolicy(metaObj, log)
 	if !reconcilePolicy.AllowsDelete() {
 		log.V(Info).Info("Bypassing delete of resource due to policy", "policy", reconcilePolicy)
-		controllerutil.RemoveFinalizer(metaObj, GenericControllerFinalizer)
+		controllerutil.RemoveFinalizer(metaObj, GenericReconcilerFinalizer)
 		log.V(Status).Info("Deleted resource")
 		return ctrl.Result{}, nil
 	}
 
 	// Check if we actually need to issue a delete
-	hasFinalizer := controllerutil.ContainsFinalizer(metaObj, GenericControllerFinalizer)
+	hasFinalizer := controllerutil.ContainsFinalizer(metaObj, GenericReconcilerFinalizer)
 	if !hasFinalizer {
 		log.Info("Deleted resource")
 		return ctrl.Result{}, nil
@@ -365,7 +234,7 @@ func (gr *GenericReconciler) delete(ctx context.Context, log logr.Logger, metaOb
 	// the finalizer
 	if (result == ctrl.Result{} && err == nil) {
 		log.V(Info).Info("Delete succeeded, removing finalizer")
-		controllerutil.RemoveFinalizer(metaObj, GenericControllerFinalizer)
+		controllerutil.RemoveFinalizer(metaObj, GenericReconcilerFinalizer)
 	}
 
 	// TODO: can't set this before the delete call right now due to how ARM resources determine if they need to issue a first delete.
